Update C++ Sample
