--- MT3620_M4_Sample_Code/FreeRTOS/MT3620_RTApp_FreeRTOS_C++/DigitalOut.h
@@ -1,16 +1,0 @@
-#pragma once
-
-extern "C" {
-#include "os_hal_gpio.h"
-}
-
-class DigitalOut
-{
-private:
-	os_hal_gpio_pin _Pin;
-
-public:
-	DigitalOut(os_hal_gpio_pin pin);
-	void Write(int value);
-
-};
