--- MT3620_M4_Sample_Code/FreeRTOS/MT3620_RTApp_FreeRTOS_C++/app_manifest.json
@@ -1,11 +1,0 @@
-{
-  "SchemaVersion": 1,
-  "Name": "MT3620_RTApp_FreeRTOS_C++",
-  "ComponentId": "40c54223-494c-430a-b7e2-c9ea406deacd",
-  "EntryPoint": "/bin/app",
-  "CmdArgs": [],
-  "Capabilities": {
-    "Gpio": [ 16 ]
-  },
-  "ApplicationType": "RealTimeCapable"
-}
