+++ MT3620_M4_Sample_Code/FreeRTOS/MT3620_RTApp_FreeRTOS_CPP/AzureSphereRTCoreToolchainVFPCpp.cmake
