+++ MT3620_M4_Sample_Code/FreeRTOS/MT3620_RTApp_FreeRTOS_CPP/CMakeLists.txt
@@ -5,22 +5,23 @@
 
 #  Copyright (c) Microsoft Corporation. All rights reserved.
 #  Licensed under the MIT License.
+
 CMAKE_MINIMUM_REQUIRED(VERSION 3.11)
 
 # Configurations
-PROJECT(MT3620_RTApp_FreeRTOS_C++ C CXX)
+PROJECT(Azure_Sphere_RTcore_FreeRTOS_CPP C CXX)
 ADD_COMPILE_DEFINITIONS(OSAI_FREERTOS)
 ADD_LINK_OPTIONS(-specs=nano.specs -specs=nosys.specs)
 
 # Executable
 ADD_EXECUTABLE(${PROJECT_NAME}
                main.cpp
-               DigitalOut.cpp
                ../OS_HAL/src/os_hal_gpio.c
                ../OS_HAL/src/os_hal_uart.c)
 
 # Include Folders
-INCLUDE_DIRECTORIES(${PROJECT_NAME} PUBLIC ./)
+INCLUDE_DIRECTORIES(${PROJECT_NAME} PUBLIC
+                    ./)
 TARGET_INCLUDE_DIRECTORIES(${PROJECT_NAME} PUBLIC
                            ../OS_HAL/inc
                            ./)
@@ -32,4 +33,4 @@ TARGET_LINK_LIBRARIES(${PROJECT_NAME} MT3620_M4_Driver)
 
 # Linker, Image
 SET_TARGET_PROPERTIES(${PROJECT_NAME} PROPERTIES LINK_DEPENDS ${CMAKE_SOURCE_DIR}/linker.ld)
-INCLUDE(${AZURE_SPHERE_MAKE_IMAGE_FILE})
+INCLUDE("${AZURE_SPHERE_MAKE_IMAGE_FILE}")
