+++ MT3620_M4_Sample_Code/FreeRTOS/MT3620_RTApp_FreeRTOS_CPP/app_manifest.json
@@ -1,0 +1,12 @@
+{
+  "SchemaVersion": 1,
+  "Name": "MT3620_RTApp_FreeRTOS_CPP",
+  "ComponentId": "BB3A83CA-41E8-4CFC-93EE-AB0AC6E7E0AB",
+  "EntryPoint": "/bin/app",
+  "CmdArgs": [],
+  "Capabilities": {
+    "Uart": [ "ISU0" ],
+    "Gpio": [ 9 ]
+  },
+  "ApplicationType": "RealTimeCapable"
+}
